Cope better with unsupported entropy driver flags
A driver may not support the particular behavior that the core wants.
Provide a reasonable path for negotiating a supported set of flags:
a driver should return `PSA_ERROR_NOT_SUPPORTED` if the core makes a
request that the driver doesn't support, and the core can try again
with fewer flags.

Invert the flag for non-blocking: make the default behavior to be blocking.
The most basic likely core just needs entropy now, so make that the most
basic behavior, with flags=0.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>