Remove MBEDTLS_NO_DEFAULT_ENTROPY_SOURCES
Remove the compilation option `MBEDTLS_NO_DEFAULT_ENTROPY_SOURCES`.
It is no longer meaningful since it's about the behavior of
`mbedtls_entropy_init()`, but this is no longer a public function.

Now the visible effect of this function is to have no entropy source at all
for the PSA random generator. This is not supported yet (but can be hacked
by enabling `MBEDTLS_PSA_CRYPTO_EXTERNAL_RNG` and providing an
always-failing callback, which is better than providing a fake entropy
source since it saves having to build a DRBG module.) So we don't need any
replacement.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>