Include limits.h where needed
This will be needed when TF-PSA-Crypto's `build_info.h` stops including
`limits.h`. It used to do so by accident because it included
`check_config.h` which wants `limits.h` to check `CHAR_BIT`. It still does
for now, but this is only temporary, until Mbed TLS is updated to include
`limits.h` where needed.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>