Change platform mutex_init to return a status code
Change the platform function for mutex initialization to return `int`. This
allows the library to know whether the initialization succeeded. (It is
possible to initialize a mutex in an unfallible way on pthread, but on
some platforms, mutex initialization is inherently fallible because it
allocates a resource.)

Keep track of whether the mutex initialization succeeded in the library
mutex structure. This will allow us to later make mutex functions safer by
not invoking platform functions on an uninitialized mutex..

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>