Remove psa_can_do_hash
The function `psa_can_do_hash()` indicates whether PSA is capable of
performing the given hash at this time. It was introduced to support hybrid
builds where an algorithm is available through both PSA and legacy
interfaces, and the PSA interface is normally preferred but the algorithm
must keep working, through the legacy interface if needs be, even during
startup when PSA is not initialized yet.

Users of client-only builds can (and must) supply their own implementation
of the function, which allows them to use md and cipher functions while
there is no connection to the crypto server. Most users are expected to have
a function that always returns 1, like TF-M does, meaning that md will only
work once communication with the crypto server has been established.

Since TF-PSA-Crypto does not support cryptography that bypasses PSA, this
can-do concept is not really meaningful anymore. We are not aware of a specific
user who needs such a hybrid scenario to work. It may be useful in theory,
but we should not commit an interface without more concrete knowledge. Thus,
remove this hook.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>