Smoke-test the release archive
Test that the library from the tar can be built in a minimal
environment (CMake, make, C compiler and a few basic shell commands). In
particular, this environment does not include Python. (This assumes that
CMake won't try to look for Python outside of `$PATH`, which does seem to be
the case in practice.)

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>