Move crypto_adjust_config_dependencies where it belongs
This is changing PSA configurations, based on conditions that are not
driver-specific, so keep it in `/include`. The file has guards based on
whether some key derivation algorithms are accelerated, but these guards are
currently irrelevant since key derivation drivers are not supported yet.

This is not meant to be included directly by applications, so move it under
`.../private`.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>