Update name of crypto_adjust_config_enable_builtins.h
Rename the file to match the pattern followed by similar files, and to
better reflect the role of the file in the context of TF-PSA-Crypto. It's no
longer about enabling legacy interfaces, since low-level crypto is mostly no
longer part of a public interface, but about enabling internal interfaces
that implement actual cryptography.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>