Reintroduce MBEDTLS_SHA3_C to guard the bulk of sha3.c
Guard the code in `sha3.c` by `MBEDTLS_SHA3_C`, as was done historically and
following the model that still exists in other files that implement
basic crypto algorithms.

Use the same macro in the corresponding test suite.

In `sha3.c`, this fixes the bug that the bulk of the code was included even
if SHA3 was accelerated (https://github.com/Mbed-TLS/TF-PSA-Crypto/issues/542).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>